Project version 1 finished

+++ src/com/cognixia/jump/model/Account.java
@@ -29,7 +29,7 @@ public class Account {
 		return balance;
 	}
 
-	public void setBalance(int balance) {
+	public void setBalance(double balance) {
 		this.balance = balance;
 	}
